Split SavedModel test.
PiperOrigin-RevId: 237898655
